Implement __nonzero__ for SpooledBytesIO, SpooledStringIO
An empty StringIO instance still has a "truthy" value. In
order to user SpooledBytesIO with the ZipFile library, the
object instance needs to be considered truthy. This patch
implements __nonzero__ so that the truthyness does not fall
back on the underlying __len__ implementation.

Added a test that verifies a SpooledBytesIO is compatible
with the ZipFile library.