Add FrozenDict type to dictutils
A FrozenDict is a dictionary whose values are
set during instance creation and are fixed from
that point on. Setting and altering values is
not allowed. This can be useful in a number of
scenarios, including setting up mapping constants
without worrying that the values will get mutated
by a misbehaving function.

There are a lot of online recipies that for creating
a FrozenDict class, but most still rely on a real
dictionary under the hood for storage. They also
tend to be extra weighty because they have an
underlying object dict.

One common solution is to use a named tuple, but this
requires setting up boilerplate for every type or
relying on factory functions. The FrozenDict utilizes
a named tuple for storage and then is further made
lighter by utilizing __slots__ to eliminate the
underlying object dict.