Use 'hex_code' instead of alias 'hex' (python3.3 fix)
The reseed function in iterutils GUIDerator uses codecs
to get hexbytes in a python2/3 compatible way. It was
relying on the 'hex' codec which an alias for 'hex_codec'.
The 'hex' to 'hex_codec' alias is missing for some reason
in Python3.3.

This patch uses the direct name of the codec instead of the
alias so that GUIDerator will work on all versions of Python,
regardless of the presence of the 'hex' alias.