Fix issue evicting duplicate keys from LRI (#155)
The insertion deque() can have duplicate items when the same key is set
multiple times. The eviction routine assumes the cache item still exists
and so can raise KeyError. This patch modifies the LRI setitem method
to use dict.pop(key) instead of del[key] so as to avoid the KeyError.

The eviction is now run in a while loop until the cache is less than or
equal to max_size since the dict.pop() eviction is not guaranteed to
actually evict anything.